Refactor PersistentResourceTest
